+++ x/stakingplus/keeper/msg_server.go
@@ -8,7 +8,6 @@ import (
 	"context"
 
 	sdk "github.com/line/lbm-sdk/types"
-	govtypes "github.com/line/lbm-sdk/x/gov/types"
 	"github.com/line/lbm-sdk/x/stakingplus"
 )
 
@@ -36,7 +35,7 @@ func (k msgServer) CreateValidator(goCtx context.Context, msg *stakingtypes.MsgC
 		if err != nil {
 			return nil, errors.ErrInvalidAddress.Wrapf("invalid grantee address: %s", msg.DelegatorAddress)
 		}
-		if err := k.fk.Accept(ctx, govtypes.ModuleName, grantee, msg); err != nil {
+		if err := k.fk.Accept(ctx, grantee, msg); err != nil {
 			return nil, err
 		}
 	}